+++ crates/astria-sequencer/CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add cache of recent execution results to mempool [#2163](https://github.com/astriaorg/astria/pull/2163).
+- Add tx result to `Executed` transaction status [#2159](https://github.com/astriaorg/astria/pull/2159).
+
 ## [3.0.0] - 2025-05-21
 
 ### Added